Move template provinces to separate file
