+++ src/provinces/province.cpp
@@ -1,4 +1,5 @@
 #include "province.hpp"
+#include "province_templates.hpp"
 #include "dcon_generated.hpp"
 #include "demographics.hpp"
 #include "nations.hpp"
@@ -9,55 +10,6 @@
 
 namespace province {
 
-template<typename T>
-auto is_overseas(sys::state const& state, T ids) {
-	auto owners = state.world.province_get_nation_from_province_ownership(ids);
-	auto owner_cap = state.world.nation_get_capital(owners);
-	return (state.world.province_get_continent(ids) != state.world.province_get_continent(owner_cap)) &&
-				 (state.world.province_get_connected_region_id(ids) != state.world.province_get_connected_region_id(owner_cap));
-}
-
-bool is_overseas(sys::state const& state, dcon::province_id ids) {
-	auto owners = state.world.province_get_nation_from_province_ownership(ids);
-	auto owner_cap = state.world.nation_get_capital(owners);
-	return (state.world.province_get_continent(ids) != state.world.province_get_continent(owner_cap)) &&
-				 (state.world.province_get_connected_region_id(ids) != state.world.province_get_connected_region_id(owner_cap));
-}
-
-template<typename F>
-void for_each_land_province(sys::state& state, F const& func) {
-	int32_t last = state.province_definitions.first_sea_province.index();
-	for(int32_t i = 0; i < last; ++i) {
-		dcon::province_id pid{dcon::province_id::value_base_t(i)};
-		func(pid);
-	}
-}
-
-template<typename F>
-void ve_for_each_land_province(sys::state& state, F const& func) {
-	int32_t last = state.province_definitions.first_sea_province.index();
-	ve::execute_serial<dcon::province_id>(uint32_t(last), func);
-}
-
-template<typename F>
-void for_each_sea_province(sys::state& state, F const& func) {
-	int32_t first = state.province_definitions.first_sea_province.index();
-	for(int32_t i = first; i < int32_t(state.world.province_size()); ++i) {
-		dcon::province_id pid{dcon::province_id::value_base_t(i)};
-		func(pid);
-	}
-}
-
-template<typename F>
-void for_each_province_in_state_instance(sys::state& state, dcon::state_instance_id s, F const& func) {
-	auto d = state.world.state_instance_get_definition(s);
-	auto o = state.world.state_instance_get_nation_from_state_ownership(s);
-	for(auto p : state.world.state_definition_get_abstract_state_membership(d)) {
-		if(p.get_province().get_nation_from_province_ownership() == o) {
-			func(p.get_province().id);
-		}
-	}
-}
 bool nations_are_adjacent(sys::state& state, dcon::nation_id a, dcon::nation_id b) {
 	auto it = state.world.get_nation_adjacency_by_nation_adjacency_pair(a, b);
 	return bool(it);
@@ -432,7 +384,7 @@ bool can_build_naval_base(sys::state& state, dcon::province_id id, dcon::nation_
 		return false;
 
 	auto si = state.world.province_get_state_membership(id);
-	
+
 	int32_t current_lvl = state.world.province_get_building_level(id, economy::province_building_type::naval_base);
 	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, economy::province_building_type::naval_base);
 	int32_t min_build = int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_naval_base));
@@ -1919,7 +1871,7 @@ std::vector<dcon::province_id> make_naval_path(sys::state& state, dcon::province
 							nearest.province.index() >= state.province_definitions.first_sea_province.index())) {
 
 
-				
+
 				if((bits & province::border::coastal_bit) == 0) { // doesn't cross coast -- i.e. is sea province
 					if(other_prov == end) {
 						fill_path_result(nearest.province);
